LayoutTests/http/tests/misc/webtiming-ssl-expected.txt

This test checks that Web Timing reports secureConnectionStart correctly and has reasonable values for connectStart and connectEnd. Note that DumpRenderTree doesn't set secureConnectionStart.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timing.connectStart is >= timing.navigationStart
FAIL timing.secureConnectionStart should be >= timing.connectStart. Was 0 (of type number).
PASS timing.connectEnd is >= timing.secureConnectionStart
PASS successfullyParsed is true

TEST COMPLETE
